requirements.txt
yfinance
pandas
duckdb
pymongo
requests
python-dotenv
pytest
pytest-cov>=4
matplotlib
mplfinance
openpyxl
pyyaml
tqdm
hypothesis
pydantic>=2.6
pydantic-settings>=2.1
dependency-injector>=4.41
httpx>=0.26
tenacity>=8.2
mongomock>=4.1
